♻️ refactor : Calendar header에 쓰이는 함수 useCalendar로 이전
