+++ src/hooks/sales/useCalendar.ts
@@ -4,12 +4,24 @@ import useSalesStore from '@/shared/store/sales';
 import { Moment } from 'moment';
 
 export const useCalendar = () => {
-  const { setCurrentDate, setIsShow, setSelectedDate, setData, setRecord } = useSalesStore();
+  const {
+    date: { currentDate },
+    setCurrentDate,
+    setIsShow,
+    setSelectedDate,
+    setData,
+    setRecord,
+  } = useSalesStore();
 
   /**
    * 날짜 클릭하면 그 날짜를 기준으로 1주일 전꺼 까지 Data를 불러오는 함수입니다.
    */
-
+  const clickPreMonthHandler = () => {
+    setCurrentDate(currentDate.clone().subtract(1, 'month'));
+  };
+  const clickNextMonthHandler = () => {
+    setCurrentDate(currentDate.clone().add(1, 'month'));
+  };
   const clickShowDataOfDateHandler = (day: Moment) => async () => {
     const { sales, formatType } = await getTodaySales(day.clone().hour(0).subtract(9, 'hour'));
     if (sales.length !== 0) {
@@ -27,5 +39,5 @@ export const useCalendar = () => {
     setSelectedDate(day.clone());
   };
 
-  return { clickShowDataOfDateHandler };
+  return { clickShowDataOfDateHandler, clickPreMonthHandler, clickNextMonthHandler };
 };
